requirements.txt
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.3
Flask-HTTPAuth==3.2.4
Flask-RESTful==0.3.7
Flask-SSLify==0.1.5
Werkzeug<0.13.0,>=0.12.0
google-api-core      
google-auth      
google-cloud-core     
google-cloud-language   
google-cloud-translate     
googleapis-common-protos
google-cloud
requests-toolbelt
requests